requirements.txt
requests
pytest
pytest-html
pytest-xdist
python-dotenv
